requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
gradio>=3.35.0
streamlit>=1.24.0
pydicom>=2.3.0
SimpleITK>=2.2.0
monai>=1.2.0
timm>=0.9.0
sentence-transformers>=2.2.0
scikit-learn>=1.2.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.14.0
seaborn>=0.12.0
albumentations>=1.3.0
kaggle>=1.5.0
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
Pillow>=9.0.0
opencv-python>=4.5.0
pyyaml>=6.0
tqdm>=4.64.0
spacy>=3.5.0
en-core-sci-sm @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_sm-0.5.1.tar.gz